Return TranslationResult instead of TCA from getTranslation() and others
Summary:
Right now when we try to obtain TCAs for translations, we either get a
nullptr or an address. If we get a nullptr, we don't know why and we
simply retry over and over at every opportunity. This is fine if the
failure is transient (write-lease for example), but wasteful if the
failure is more permanent (TC is full).

We want to be smarter about these scenarios. As a first step, we need
to know how long the failure is going to persist. Instead of returning
a TCA, return a new TranslationResult struct. This encapsulates either
a TCA (representing success), or a "scope" (representing failure). The
scope is an enum describing how long the failure is expected to
persist. Right now, they are transient, request, and process.

Re-organize some of the "should translate" checks to be in a smarter
order. In general, we want to do the "process", then "request", then
"transient" checks in that order (to catch permanent failures in
priority over others). Add a per-request cached boolean to latch JIT
time-out failures.

Reviewed By: mofarrell

Differential Revision: D25761261

fbshipit-source-id: 91ba035ae633e4a1c89f676ea968213c7e245a16