Use "no translate" version of resumeHelper when appropriate
Summary:
Add a new version of resumeHelper (called resumeHelperNoTranslate)
which does not call getTranslate and only attempts to use existing
translations (and interps otherwise). Likewise add a matching set of
unique stubs which call resumeHelperNoTranslate like the set for
resumeHelper. If a call to getTranslate returns Scope::Request or
Scope::Process, use the appropriate stubs for the "no translate"
variant. This means that once we know we can't generate any new
translations for this request, we'll stop calling getTranslation over
and over, which is more efficient.

Note that this diff does not attempt to smash translations or anything
like that, so we will continue to call handleServiceRequest and
family.

Differential Revision: D25769820

fbshipit-source-id: 020960aa58bdd919b189d17b2599d1061cfbd7c2