Add Vinstr::loadzwq
Summary:
MonotypeDict's tombstones field is only 2 bytes, so we want an efficient loadzwq instruction. In addition to compiling to a single load on its own, this instruction plays well with other vasm optimizations (e.g. a loadzwq/cmpqi can be simplified to a cmpwi).

On ARM and PPC, the new instruction is equivalent to loadw, but on x86-64, we need to expose a new machine-specific instruction too.

Reviewed By: oulgen

Differential Revision: D25836214

fbshipit-source-id: 082b99402059e8aa0f2bb4cbcec58650c756cf45