Lift the MonotypeDict capacity limit
Summary:
One way that MonotypeDicts can escalate in prod is because we hit the maximum size limit, 0xc000. It's hard to defend against this issue at a sampling rate of 1/1000, because array sizes have a long tail.

We fix the issue by expanding MonotypeDict::Index to 4 bytes instead of 2, which is sufficient, since Hack arrays have a maximum size of 1 << 31. Of course, the trick is to do so without losing too much capacity. It turns out that we had some slack because the array header has a fixed size. We only have to drop exactly one element from the capacity of MonotypeDict at each size, such that increasing sizes follow the following progression:

| Doubling | Size Index | Num Elms | Num Indices
| 0 | 7 | 5 | 8
| 1 | 11 | 11 | 16
| 2 | 15 | 23 | 32
| 3 | 19 | 47 | 64
| 4 | 23 | 95 | 128
| 5 | 27 | 191 | 256
| ... | ... | ... | ...

The number of elements is 6 * 2^n - 1, just one less than the old capacity.

The tricky part with this change is that we still only have a 2-byte "used" fields. In the previous diff, we changed this to be a "num tombstones" field. That means that we need to shrink the array when we hit the max tombstone count. We take this opportunity to shrink the array when the load factor drops, too, and we write the compaction helper carefully to avoid generating a lot of code.

This change is quite tricky, and hits a lot of edge cases that aren't tested. We write a new unit test that robustly hits all remove cases by adding 2^n elements to a dict and then removing them in a "cache-destroying permutation" order - a Hamiltonian path through a cube. The test revealed tons of refcounting bugs. This diff looks way too simple for how much pain it caused me.

Reviewed By: ricklavoie

Differential Revision: D25843223

fbshipit-source-id: d7483e8b9673087b55a3a33f1ed74188ee45d06c