Partial vtables for MonotypeVec
Summary:
In order to reduce dynamic dispatch, we introduce "partial vtables" for abstract bespoke arrays. While concrete layouts must provide an implementation of every layout function, abstract layouts may now provide implementations for a subset of layout functions. If, while lowering a bespoke IR operation, the known abstract layout of the bespoke array offers a layout function, a direct call will be emitted instead of dispatching to BespokeArray::Fn.

For MonotypeVec, we generalize the MonotypeVec layout function implementations to also handle the EmptyMonotypeVec case (for which we have ensured `type() == KindOfUninit`) and use these to populate the partial vtable for EmptyOrMonotypeVec and TopMonotypeVec.

A forthcoming diff will enable partial vtables for MonotypeDict.

Differential Revision: D25786609

fbshipit-source-id: 129b7f58355cd46997182e62d8dfcf60322620b1