Add missing allocation failure check in ext_gd.cpp
Summary:
There was a missing allocation failure check in one of the
places we were allocating memory for the Thumbnail. This meant the
buffer was a nullptr, so when we attempted to free it a few lines
below, we could crash (we'd do pointer arithmatic on the pointer and
wrap around to some big number).

Do the standard allocation failure check.

Reviewed By: oulgen

Differential Revision: D25844741

fbshipit-source-id: 5f9cda54b9295215eca0a6824818822ea638d9cf