Provide ocaml_ffi_with_arena macro
Summary:
It is currently not possible to use arena-allocated types (e.g., the types in the `oxidized_by_ref` crate) with the `ocaml_ffi` macro because such types cannot implement `FromOcamlRep`. These types implement `FromOcamlRepIn` instead, which allocates the Rust values in a given arena.

Since we cannot use `ocaml_ffi` for these types, we have to write out the full unsafe FFI function declaration by hand (as in rust_decl_ffi.rs before this diff), providing opportunities to make mistakes. It would be more convenient if we had a version of `ocaml_ffi` which supported arena-allocated values.

This diff adds the `ocaml_ffi_with_arena` macro for this purpose. It expands into code which constructs an arena, uses `FromOcamlRepIn` to convert the parameters into Rust values allocated in that arena, runs the inner logic of the function, and converts the returned value to an OCaml value. This diff also uses the new macro in rust_decl_ffi, to replace the handwritten equivalent.

Reviewed By: shiqicao

Differential Revision: D25811736

fbshipit-source-id: 1b035555734612c77ef60a8befc329741d8d955c