Remove ocaml_ffi_no_panic macro
Summary:
The ocaml_ffi_no_panic macro behaves like ocaml_ffi, except it does not introduce a panic boundary with `catch_unwind`. Instead, it trusts the user to ensure that their function body will never raise a panic. I wrote this macro specifically for rust_facts_ffi, because rust_facts_ffi did not previously use catch_unwind, and I was concerned that using it might introduce a performance issue.

I think this was the wrong call. `catch_unwind` is not as expensive as I had thought, and failing to catch a panic in an FFI function causes undefined behavior (the stack-unwinder will continue unwinding into OCaml stack frames). We shouldn't provide a convenience macro that can trigger UB--we should just use `ocaml_ffi` instead. If this causes a perf problem in the facts parser, we can revisit the issue (but any solution should not involve providing this macro in `ocamlrep_ocamlpool`, because it isn't safe). I don't expect to see any perf regression as a result of this change, though.

Reviewed By: shiqicao

Differential Revision: D25814407

fbshipit-source-id: 7799ed623fec8f50f6274ea3f604dc3b272e279c