Clean up HH\Contexts in naming special names
Summary:
Everything fully qualified in Naming_special_names modules
that is in namespace HH begins with `\` (a single backslash), therefore
fix the constants `Coeffects.(unsafe_)contexts`.  This allows those
special names to be used in other logic, such as namespace lookup (coming soon).

In order to avoid perf regression, cache the constructed namespaces in which
contexts are elaborated.

Reviewed By: DavidSnider

Differential Revision: D25858798

fbshipit-source-id: a66ba47c0fad524f4b15ee36426920c1251d5faa