Respect bespoke coloring in live translations
Summary: In live translations, we don't want to specialize bespoke sinks, because one of the main reasons we might create such translations is because we failed a guard on a sink. However, we should still specialize bespoke sources. Not doing so can cause a cascading sequence of guard failures. See also: https://en.wikipedia.org/wiki/Robustness_principle

Reviewed By: colavitam

Differential Revision: D25860466

fbshipit-source-id: abab1e549b53096c32f13b8470056d7d3dce8271