Immutable locals: prevent reassignment of parameters used in path-dependent context
Summary:
If `$x::C` appears in a context, prevent reassignment of parameter `$x`.

Functionally, this diff adds immutable locals to Hack. We always bind locals with `Env.set_local`. Each local has a 63 bit integer id attached to it. This diff adds an optional parameter to `Env.set_local` which sets the high bit. Then, when we try to change the expression ID in `Env.set_local_expr_id`, if the immutable bit is high, we raise an error.

Reviewed By: losvald

Differential Revision: D25795454

fbshipit-source-id: 000c57fd0f9a25f51e5f587d06da4dbfbae67c49