UX: ban by default non-contexts in context list
Summary:
This isn't strictly needed in the long-term; it is a
short-term artifact of the `defaults`=`nothing`.

Normally, functions with contexts that include:

- primitive types (e.g., `int`)
- user-defined types

won't be callable because the default context won't provide any capabilities that are subtype of, for instance, `int`.

Currently, though, we need to be careful not to allow such contexts,
because if users starts to write `int` as contexts, for example,
this won't fail (as `int <: nothing`) and
their code won't be migratable after we switch to `defaults != nothing`.

Also ban generic as context, not to (yet) expose internal usage of generics for effect polymorphism.

Since the current capability lattice isn't as complex as it will be in the near future,
don't ban such contexts in test cases to keep the increased coverage, i.e.,
allow for testing on a much complex, provisional capability lattice.
Similarly, gate the case of generics so that we can ungate it in tests and test internal implementation details.

Reviewed By: DavidSnider

Differential Revision: D25858844

fbshipit-source-id: c37ebf0b68b446ebfcb0e9e6941aced5be006bd3