Write a program to verify saved state
Summary:
Compare a depgraph `l`, against another `r` and report on the edges in `r` not in `l` (missing edges) and the edges in `l` not in `r` (extra edges).

The idea is to use something like this in CI to detect incremental saved states with depgraph corruption.

Next steps:
- Support for the 64 bit depgraph work;
- Decoding missing edges so that we can report in terms of source symbols instead of their hashes;
- I'm open to entertaining feature requests!

There's lots to do; this diff represents a beginning not an end!

Reviewed By: ljw1004

Differential Revision: D25793095

fbshipit-source-id: b6d7e23677b87e761ec669368823caefbc88edbe