Remove comment at the end of error variants
Summary:
Now that error numbers are spelled out on each variant, I think it's obvious what one needs to do to add an error code.

Every time we add a variant, we move the comment down one, so the blame gets destroyed on the prior error code. If you put it on a newline after all the variants, ocamlformat puts it back. I don't want to do a one-off ocamlformat config for this one thing.

Reviewed By: losvald, Wilfred

Differential Revision: D25873694

fbshipit-source-id: b589da175fc0adfdcbcbdc60aea2d0dede7cc6e2