Add type inference functions for mutations
Summary:
For append/remove/set ops, given the layout of the input and types for the keys and vals, we may be able to infer a nontrivial layout for the output. We add functions for all three mutations, even though we only have a place to use the append and set ops right now - removes are always generic, for now. There were some tricky bugs in earlier versions of this diff:
1. When we append or set a value of the wrong type to EmptyOrMonotypeVec or EmptyOrMonotypeDict, we must escalate to Top - the array could be empty, in which case it would stay bespoke. But MonotypeVec and MonotypeDict can escalate to Vanilla instead. I pulled out the non-empty logic into a helper and escalated vanilla -> top separately.
2. There may not be a layout corresponding to the given (KeyType, DataType) pair - for instance, there's no EmptyOrMonotypeDict layout for a persistent-flavor type, and there's no MonotypeVec layout for Uninit.

Reviewed By: colavitam

Differential Revision: D25870949

fbshipit-source-id: d6e408b56f2c9ba4f3890fab3e63123074e16336