Ignore missing nodes in FFP JSON tests
Summary:
allow-large-files So this is a diff that deletes over 60000 lines and reduces the size of the codebase needed; not really sure if there's any way to get around it and the files are already there.

This diff adds an option --ignore-missing-json that gets rid of missing nodes in the JSON output of hh_parse. It is only used for FFP Json tests, where missing nodes really don't help for what we'd want to test for and add a lot of extra noise when a new feature is being built.

{F359823399}

The only file worth reviewing is generate_full_fidelity_data.ml (and hh_parse.ml), everything else is generated.

Now, when we make a change that adds a new feature to the FFP, the tests act as a regression test to make sure old parsing/tests did not change with the exception of missing nodes.

Reviewed By: oulgen, Wilfred

Differential Revision: D25916261

fbshipit-source-id: 31679f25a53018ae98d467060c9225ca824825c1