Devirtualize destructors for vanilla-only arrays
Summary: We never use bespoke keysets, so we don't want to pay the cost of layout checks for these array-likes. By the same token, we should use the vanilla destructor for them. This diff won't matter much in our current configuration, but we want to run monovec-only and monodict-only tests, for which it'll be more important.

Reviewed By: colavitam

Differential Revision: D25948525

fbshipit-source-id: e76d64bc2e68ede65077b4a37bd12dfd056a4e7e