Fix another bug with type-limited bespokes
Summary: perf data from the prior diff led me to this bug: even if we claim that a given array type will never be bespoke, we fail to specialize iterators for it. Specifically, I saw a large increase in the cost of iterNextKArrayMixedPointer, etc. on the diff that should have restricted bespokes to vec/varray. I've checked that we are now able to specialize dict iterators in this case.

Reviewed By: ricklavoie

Differential Revision: D25949307

fbshipit-source-id: c3d44363a4cd42085c8418a69b22b6f2009cd945