Refactor two flags to sub-typing
Summary: Sub-typing had two boolean flags `allow_subtype_of_dynamic` and `treat_dynamic_as_bottom`. Both were somewhat (but not entirely) descriptive of how they changed the behaviour of sub-typing, but it was not entirely clear why/when they should be used. In fact they are only used for coercion from/to type dynamic, and should never both be true at the same time. This refactors into a single flag `coerce` that has optionally the direction of coercion (where None means that no coercion is happening).

Reviewed By: vsiles

Differential Revision: D25947091

fbshipit-source-id: d37b902add5a7510dd56ffc579794a2999875d52