Fix non-terminating checking with intersection type bound on abstract type constant; instantiate this-type with intersection type of receiver instead of splitting
Summary:
This diff combines two major fixes. Unfortunately they can't be implemented as separate diffs as they interact.

**(1) Fix non-terminating checking with intersection type bound on abstract type constant**

Hack loops when checking the following code:
```
interface K {
  abstract const type TC as K;
}

interface J {
  abstract const type TC as (K & this);
  public function get(): this::TC;
}

function testit(J $x): void {
  $y = $x->get();
  $y->get();
  $y->get();
```
(Although this uses the experimental intersection type syntax, similar looping behaviour can be induced in other ways.)

The fix is to change `get_concrete_supertypes` so that it properly decomposes intersection types. Then we reconstruct an intersection type when checking function calls, instead of using `try_over_concrete_supertypes`, which is now redundant.

**(2) Instantiate this-type with intersection type of receiver instead of splitting**

Consider this code:
```
class C {
  public function getit():this { return $this; }
}

function testit<T>(T $x):T {
  $x as C;
  return $x->getit();
}
```
This should pass the type-checker. The call to `getit()` should preserve the type: `$x` has type `T & C` due to the refinement, and the signature of `getit` essentially says that the result type is the type of the receiver.

Unfortunately, Hack currently "loses" the fact that the receiver type is `T & C` and not just `C`. That's because it deals with an intersection type for the receiver by attempting to extract a method type for each conjunct in turn, instantiating `this` with the conjunct, and then forming a conjunction of function types from those *that succeeded*. In this example, it fails to find a method in `T`, but does find one in `C`, instantiating `this` with `C` (in fact, with an expression-dependent type whose bound is `C`). The end result is that `T` is lost.

The actual mechanics of instantiating the `this` type are somewhat complex, involving passing a continuation function (`k_lhs`) to `obj_get` that gets applied "on demand". In this diff, we instead just pass down the `this_ty` type from the top, *before* we split apart the intersection type into conjuncts.

As part of this diff, we improve the `obj_get` interface by using named parameters.

Reviewed By: CatherineGasnier

Differential Revision: D25886508

fbshipit-source-id: 818877dab08ea984f77ef05bcf624141f2fc4be3