Add dumpcoverage mode and get_executable_lines function to hhvm
Summary:
These APIs will provide a way to query an upper-bound on the lines that may be returned during execution of a file with code coverage enabled. No guarantees are made about that all of these lines must be executable (and in particular optimizations or changes to the bytecode may change the nature of which lines are considered executed in ways that don't always match this API).

Here we use the line table so this should at the very least always be an upper-bound.

Reviewed By: ricklavoie

Differential Revision: D25960620

fbshipit-source-id: a2ee5bf04e6f1a299c8f75c34f2da310117a0737