Don't check translation counts until lease is held
Summary:
We check that we don't overflow the max translation per source record while
acquiring permission to translate.  We did this twice once before grabbing the
lock, and once after grabbing the lock.  However doing this before grabbing the
lock can be unsafe.  If the ProfileMaxTranslations > MaxTranslations we can
observe issues in non retranslateAll mode.

Thread A:
  - Enters `retranslateOpt` (to generate an optimized retranslation)
  - Grabs the write lease
  - sets the func as having been optimized
Thread B:
  - Enters `retranslate` (to generate another translation for types we have not
    seen)
  - Function is marked as optimized so the retranslate request has its type set
    to Live (as opposed to Profile which has the higher max translation limit)
  - Goes to acquire the lease, but before acquiring checks the srcRec has fewer
    then max translations, which fails leading to a process crash.

With this change, thread B will not perform that last check, and instead
will fail to acquire the lease.  This will give time for thread A to complete
optimized retranslation and remove all profile translations from the srcRec.

Reviewed By: ricklavoie

Differential Revision: D25954090

fbshipit-source-id: 03ed214d6e551a4769a3c22f364a79864b008bb5