Elaborate type arguments for ctx in standard namespace
Summary:
This was breaking things like `cipp_of<A>`, which was elaborating to `\HH\Contexts\cipp_of<\HH\Contexts\A>`. I had to unroll a bit of the logic and basically change the `contexts` elaboration not to use the on_hint method from super.

I was hoping there would be a way to override `method! on_contexts` with an optional argument to set the namespace and make it final, but I don't think this is supported in OCaml.

Reviewed By: losvald

Differential Revision: D25931248

fbshipit-source-id: aa942a8e2586338ffddacca0196d76e4916d5618