Preemptively update tests for UndefinedVariableException 12/n
Summary:
From summary 1/n:
- Would like to have UndefinedVariableException fatal in HHVM instead of thrown in www.
- There are ~500 tests, so in order to make code review possible, I'm breaking up tests before landing that change

These tests are checking an array of inputs on different functions. Since undefined variable will always throw an exception, it seem redundant to pull these arguments out of the array into a try catch for each test, and wouldn't test the function these tests are testing. So for these, we should remove undefined/unset args from the array of inputs to be tested.

- These changes were largely generated with sed, and are nearly all the same

Reviewed By: ricklavoie

Differential Revision: D25966576

fbshipit-source-id: c2f862418415f07d482fb0fedc797e74a69e790f