Long-lived workers
Summary:
This diff adds a new mode where multiprocessing workers are not killed at the end of each job. This mode is enabled by setting `use_worker_clones=false` in `hh.conf`. By default, the new mode is disabled (i.e., clones are used).

Big picture: The new mode is threaded from the config down to `WorkerController.make`, there it is passed to the worker process as a boolean parameter to the worker entry point. In the worker entry point (a closure in `WorkerController.register_entry_point`), the boolean is used to decide which of `unix_worker_main` or `unix_worker_main_no_clone` should run.

Some non-trivial changes were made to implement the cancellation logic in the new mode.

- *Before* - A worker is asked to cancel by setting its "should exit" field in shared memory to 1. In most operations of `hh_shared.c` this bit is tested and if it is set, the Worker_should_exit exception is thrown. The exception is quickly handled in the OCaml bindings for the C functions. This handling consists in first, executing an "on worker cancelled" hook and second, exiting with code 0.
- *After* - Having the heap operations exit would kill long-lived workers. The solution I chose to avoid this problem is to stop handling Worker_should_exit deep in the call stacks and instead let it bubble up to the workers' entry point. If the exception surfaces and we are not in long-lived mode we exit the clone, but if we are in long-lived mode we simply wait for the next job.
  **Nota Bene:** this change affects both the long-lived worker mode *and* the regular mode which we will be using in production.

Cancellation in long-lived mode could have been handled by respawning the cancelled worker processes. I have not tried this approach so I cannot really tell how hard it would be to implement. The advantage of the approach in this diff is that the invariant of WorkerController is not broken (Unix workers live forever). The advantage of respawning is that cancellation has a better latency (the exception can be inadvertently caught by a `try ... with _ -> ...`).

Reviewed By: andrewjkennedy, ljw1004

Differential Revision: D25615243

fbshipit-source-id: cdaab184e0dc0e395e1c2b2ae16f7997ac789131