Avoid filename comparisons during position merging
Summary:
I observed (using `perf`) that we are spending a lot of time comparing paths during decl-parsing:

```
     1.73%          5517  bench            bench               [.] core::iter::traits::iterator::Iterator::eq
            |
             --1.52%--core::iter::traits::iterator::Iterator::eq_by (inlined)
                       core::iter::traits::iterator::Iterator::eq
                       <std::path::Path as core::cmp::PartialEq>::eq
                       |
                        --1.52%--<std::path::Path as core::cmp::PartialEq>::ne (inlined)
                                  <&std::path::Path as core::cmp::PartialEq>::ne (inlined)
                                  <core::option::Option<&std::path::Path> as core::cmp::PartialEq>::ne (inlined)
                                  <oxidized_by_ref::manual::relative_path::RelativePath as core::cmp::PartialEq>::ne (inlined)
                                  <&oxidized_by_ref::manual::relative_path::RelativePath as core::cmp::PartialEq>::ne (inlined)
                                  <oxidized_by_ref::manual::pos::Pos>::merge
                                  <direct_decl_smart_constructors::direct_decl_smart_constructors_generated::DirectDeclSmartConstructors>::merge::<&oxidized_by_ref::manual::pos::Pos, &oxidized_by_ref::manual::pos::Pos>
                                  |
                                   --0.66%--<direct_decl_smart_constructors::direct_decl_smart_constructors_generated::DirectDeclSmartConstructors>::merge_positions (inlined)
```

This is because `Pos::merge` checks that both positions being merged are in the same file, and Rust does not emit pointer-equality checks for comparisons of references.

We can avoid this by not performing the comparison in the direct decl parser at all--there is no need, since all positions we handle are guaranteed to be in the same file. We can also improve performance in other situations by manually adding reference equality checks for filenames.

There's no need to make these changes in the `oxidized` version of Pos (as opposed to `oxidized_by_ref`, which is changed in this diff) because those positions wrap their filenames in an Rc, and Rc *does* perform a reference equality check in its implementation of PartialEq.

Reviewed By: shiqicao

Differential Revision: D26003505

fbshipit-source-id: 0fe87bf483dc92d328caa4a0033fcf383454177f