Do not compute number of RHS bits dynamically
Summary:
We recently added the Tiny variant of Pos, which represents a span (i.e., a start offset/line/column and an end offset/line/column) as several bitfields packed into a single integer. We currently represent these fields with a runtime value of type Field. When computing the number of bits to the right of a field, we iterate over the variants of Field, summing the number of bits in each. This operation is done each time we construct a Tiny pos or read one of its fields, and it adds up to a considerable amount of time in the direct decl parser. `perf` recorded the following in a benchmark of the direct decl parser:

```
     2.61%          8312  bench            bench               [.] <core::iter::adapters::Map<core::iter::adapters::Skip<core::iter::adapters::SkipWhile<oxidized::manual::pos_span_tiny::FieldIter, <oxidized::manual::pos_span_tiny::Field>::n_rhs_bits::{closure#0}>>, <oxidized::manual::pos_span_tiny::Field>::nbits> as core::iter::traits::iterator::Iterator>::sum::<usize>
            |
            |--1.12%--<core::iter::adapters::Skip<core::iter::adapters::SkipWhile<oxidized::manual::pos_span_tiny::FieldIter, <oxidized::manual::pos_span_tiny::Field>::n_rhs_bits::{closure#0}>> as core::iter::traits::iterator::Iterator>::fold::<usize, core::iter::adapters::map_fold<oxidized::manual::pos_span_tiny::Field, usize, usize, <oxidized::manual::pos_span_tiny::Field>::nbits, <usize as core::ops::arith::Add>::add>::{closure#0}> (inlined)
            |          <core::iter::adapters::Map<core::iter::adapters::Skip<core::iter::adapters::SkipWhile<oxidized::manual::pos_span_tiny::FieldIter, <oxidized::manual::pos_span_tiny::Field>::n_rhs_bits::{closure#0}>>, <oxidized::manual::pos_span_tiny::Field>::nbits> as core::iter::traits::iterator::Iterator>::fold::<usize, <usize as core::ops::arith::Add>::add> (inlined)
            |          <usize as core::iter::traits::accum::Sum>::sum::<core::iter::adapters::Map<core::iter::adapters::Skip<core::iter::adapters::SkipWhile<oxidized::manual::pos_span_tiny::FieldIter, <oxidized::manual::pos_span_tiny::Field>::n_rhs_bits::{closure#0}>>, <oxidized::manual::pos_span_tiny::Field>::nbits>> (inlined)
            |          <core::iter::adapters::Map<core::iter::adapters::Skip<core::iter::adapters::SkipWhile<oxidized::manual::pos_span_tiny::FieldIter, <oxidized::manual::pos_span_tiny::Field>::n_rhs_bits::{closure#0}>>, <oxidized::manual::pos_span_tiny::Field>::nbits> as core::iter::traits::iterator::Iterator>::sum::<usize>
            |          <oxidized::manual::pos_span_tiny::Field>::n_rhs_bits (inlined)
            |          <oxidized::manual::pos_span_tiny::PosSpanTiny>::project (inlined)
            |
             --0.72%--<oxidized::manual::pos_span_tiny::FieldIter as core::iter::traits::iterator::Iterator>::nth (inlined)
                       <oxidized::manual::pos_span_tiny::FieldIter as core::iter::traits::iterator::Iterator>::next (inlined)
```

We can avoid doing this work repeatedly if we compute the number of bits to the right of each field at compile time (in Rust, at least--if OCaml doesn't fold these constants, I think the work will be done once at module-load time instead). This diff does so.

This diff also addresses a case that the previous implementation would not handle correctly if we were to allocate more than 0 bits for the line-number-increment and beginning-of-line-increment. Previously, we stored a column-increment--the difference between the end column number and the start column number. In multi-line spans, however, we do not have the guarantee that the end column is greater than or equal to the start column. If we want to store the difference between the end and start columns, the field would need to be a signed integer. Instead, this diff replaces the column-increment field with a byte-width field, representing the difference between the end character number and the start character number.

Reviewed By: CatherineGasnier

Differential Revision: D25908793

fbshipit-source-id: 399f0d6d230dde5530b6bdfe07da3963e80fd130