Revise namespaces_rust to be usable in direct decl parser
Summary:
It's not easy for us to use the `oxidized::namespace_env::Env` type in the direct decl parser. It's expensive to clone, and introduces borrowing struggles (the direct decl parser deals in `&'a str`s, where `'a` is the lifetime of the source text and arena, but the oxidized namespace env just stores `String`s).

We could use the `oxidized_by_ref::namespace_env::Env` type instead, but if we'd like both env types to share namespace elaboration logic, we need to abstract over whether it is the `oxidized` one or the one from `oxidized_by_ref`. This diff does so.

Reviewed By: shiqicao

Differential Revision: D26003508

fbshipit-source-id: 803d4dd1360d5fdc498a1d8208c0e3378c56ec3b