Begin naming table comparison
Summary:
A quick binary that takes two naming tables from saved state and performs a diff on them.

There are two formats that the naming table can be saved as in the naming table, the legacy marshaled ocaml blob and the more updated sqlite table. We attempt to handle diffing between both of these formats.

```
/path/to/diff_naming_table $path_to_first_naming_table $path_to_second_naming_table
```

The default assumes that both are marshaled ocaml blobs.
You can use `--f1-sqlite` and `--f2-sqlite` to signal that the first or second files are meant to be paths to sqlite files rather than marshaled ocaml blobs.

The result is a printed list of changed files. If there are alot of changes, you can use the `--count` option to simply print a count of each type of change: Added/removed files, changed files, added/removed error files.

Reviewed By: shayne-fletcher

Differential Revision: D25796169

fbshipit-source-id: 00dc673fd48ebe6c3de60b060600ae603c48c152