Add a surprise check on Jmp / JmpZ / JmpNZ
Summary:
When these bytecodes are used for back-edges (e.g. for loops and while loops), we should do a surprise check before taking the jump, so that requests caught in infinite loops will time out.

JmpNS means "jump, no surprise", so we shouldn't do it in that case. We use it for edges that are guaranteed not to be back-edges.

Reviewed By: mofarrell

Differential Revision: D26036689

fbshipit-source-id: be3fef2521fa388333e7f182ab10409d683dc9bf