Desugar to a function call instead of `ExprTree`
Summary:
Previously, we desugared an Expression Tree to the `ExprTree` class.

```
Example`4`;

->

... return new ExprTree(...) ...
```

This diff changes the desugaring to now be:

```
Example`4`;

->

... return Example::makeTree(...) ...
```

This allows the framework to decide what class is ultimately instantiated from the Expression Tree. Additionally, since we made `Spliceable` separate from `ExprTree`, this change also doesn't impact the ability to splice various things together nor is `ExprTree` a keyword that Hack needs to pay attention to.

We change the tests now:

```
  public static function makeTree<TVisitor, TResult, TInfer>(
    ?ExprPos $pos,
    string $filepath,
    dict<string, mixed> $spliced_values,
    (function(TVisitor): TResult) $ast,
    (function(): TInfer) $err,
  ): ExprTree<TVisitor, Code::TAst, TInfer> {
    ...
  }
```

Everything remains the same, just added a layer of indirection here.

Reviewed By: Wilfred

Differential Revision: D25934518

fbshipit-source-id: 8904987cddb189f5eb137dc86acc72dadc936aee