Fix slow dependency-hash-to-filename table building
Summary:
Previously, loading the dependency-hash-to-filename table was extremely slow for the 64-bit dependency graph. For 32-bit it took ~26s, for 64-bit a whopping ~73s!

I did some profiling, and apparently the difference was entirely due to allocating dependency sets:
- In 32-bit mode, we use OCaml's `ISet`, which allocates fast thanks to OCaml fast allocation mechanism
- In 64-bit mode, we use Rust's `OrdSet`, which requires a heap allocation from the moment you insert an element. Malloc/free is a lot slower than OCaml's allocator.

Luckily, we don't need sets here, as all hashes are expected to be unique; and even if there are duplicates it shouldn't be a problem as the operation performed on them is idempotent.

So, to avoid Rust's allocator, we just use standard OCaml lists to collect the hashes.

*Side note: avoiding unions improves 32-bit performance as well by a tiny bit.*

Reviewed By: CatherineGasnier

Differential Revision: D26073842

fbshipit-source-id: d1c355be56d6f49e0711698852e2d18e1afae00e