Dynamically generate bitmaps for co-effects encoding
Summary:
This diff converts encoding of co-effects in the runtime from an explicitly listed enum to dynamically generated bitfield. With this new encoding, we are able to express multitude of new configurations including `A, B` and `A & B` for any co-effect `A` and co-effect `B`.

When the user specifies a co-effect in `.static_coeffects` hhas directive, we convert this co-effect into a coordinate system of capabilities. This mapping is available as `CoeffectsConfig::s_coeffects_to_capabilities`.

During startup based on graphs of all capabilities given in a coeffects config file, we generate encodings for all capabilities. At the time of this diff, the current encodings are as follows:
```
rx_defaults   : 0000000000000000
rx_local      : 0000000000000001
rx_shallow    : 0000000000000010
rx            : 0000000000000011
rx_pure       : 0000000000000111
cipp_defaults : 0000000000000000
cipp_local    : 0000000000001000
cipp_shallow  : 0000000000010000
cipp          : 0000000000011000
cipp_global   : 0000000000111000
cipp_pure     : 0000000001111000
```
Note that the above is generated dynamically and doesn't need to remain as is.

Reviewed By: jano

Differential Revision: D25857177

fbshipit-source-id: 383f87affa13b7c7a255bb54132891a274f6758e